+++ cogs/install.scm
@@ -13,6 +13,7 @@
 (define (append-with-separator path)
   (if (ends-with? path "/") (string-append path "cogs") (string-append path "/cogs")))
 
+;; Should make this lazy?
 (define *STEEL_HOME* (~> "STEEL_HOME" (env-var) (unwrap-ok) (append-with-separator)))
 
 (define/contract (parse-cog module)
