+++ crates/example-dylib/Cargo.toml
@@ -11,3 +11,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 steel-core = { path = "../steel-core", version = "0.2.0", features = ["modules"] }
+abi_stable = "0.11.1"
