+++ crates/steel-core/src/lib.rs
@@ -18,6 +18,7 @@ mod conversions;
 // pub mod jit;
 pub mod parser;
 pub mod steel_vm;
+
 #[cfg(test)]
 mod tests;
 pub(crate) mod values;
