--- crates/steel-core/src/steel_vm/threads.rs
@@ -1,14 +1,0 @@
-use crate::values::functions;
-use crate::{rvals::SteelVal, values::functions::ByteCodeLambda};
-
-// fn can_be_copied_to_thread(function: &ByteCodeLambda) {
-//     function
-// }
-
-// fn deep_clone(val: SteelVal) {
-//     if let SteelVal::BoxedFunction(b) = val {
-
-//     }
-// }
-
-// Can be naively translated to another thread?
