+++ crates/steel-toml/Cargo.toml
@@ -12,3 +12,4 @@ crate-type = ["cdylib"]
 [dependencies]
 steel-core = { path = "../steel-core", version = "0.2.0", features = ["modules"] }
 toml = "0.7.1"
+abi_stable = "0.11.1"
